+++ imap_tools/__init__.py
@@ -1,7 +1,7 @@
 from .query import AND, OR, NOT, Header, UidRange, A, O, N, H, U
 from .mailbox import BaseMailBox, MailBox, MailBoxUnencrypted
-from .message import MailMessage, Attachment, MailMessageFlags
+from .message import MailMessage, MailAttachment, MailMessageFlags
 from .folder import MailBoxFolderManager, MailBoxFolderStatusOptions
 from .errors import *
 
-__version__ = '0.34.0'
+__version__ = '0.35.0'